fixes for transition from int to long

+++ test/common.hpp
@@ -15,15 +15,15 @@ public:
 };
 class Complex : public gorage::Json {
 public:
-  int i;
+  long i;
   double d;
   std::string s;
 
   Complex() {}
-  Complex(const int &i, const double &d, const std::string &s) : i(i), d(d), s(s) {}
+  Complex(const long &i, const double &d, const std::string &s) : i(i), d(d), s(s) {}
   Complex(const Structure &structure) {
     const auto dict = cast<Dict>(structure.value());
-    i = get<int>(dict, "i");
+    i = get<long>(dict, "i");
     d = get<double>(dict, "d");
     s = get<String>(dict, "s").s;
   }
